requirements.txt
streamlit
pandas
plotly
python-pptx
openpyxl